requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.21.0
transformers>=4.21.0
accelerate>=0.20.0
opencv-python>=4.5.0
pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.64.0
tensorboard>=2.9.0
omegaconf>=2.2.0
einops>=0.6.0
datasets>=2.10.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
wandb>=0.13.0